WIP added sass and rearranged file structure
