+++ todo/src/css/styles.css
@@ -1,0 +1,13 @@
+.App {
+  text-align: center; }
+  .App .header {
+    background-color: #282c34;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: flex-start;
+    font-size: calc(10px + 2vmin);
+    color: white; }
+
+/*# sourceMappingURL=styles.css.map */
