+++ todo/src/css/styles.scss
@@ -1,0 +1,15 @@
+$default-padding: 2rem;
+
+.App {
+  text-align: center;
+  header {
+    background-color: #282c34;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: flex-start;
+    font-size: calc(10px + 2vmin);
+    color: white;
+  }
+}
