+++ todo/src/reducers/Reducer.js
@@ -1,14 +1,14 @@
-import React, { useReducer } from 'react';
+// import React, { useReducer } from 'react';
 
-const initialState = [
-  {
-    task: 'Finish this to-do app...',
-    completed: false,
-    id: 1,
-  },
-];
+// const initialState = [
+//   {
+//     task: 'Finish this to-do app...',
+//     completed: false,
+//     id: 1,
+//   },
+// ];
 
-function reducer(state, action) {
+export function reducer(state, action) {
   switch (action.type) {
     case 'ADD_TODO':
       return Object.assign({}, state, {
@@ -23,31 +23,29 @@ function reducer(state, action) {
   }
 }
 
-const ToDo = () => {
-  const [state, dispatch] = useReducer(reducer, initialState);
+// const ToDo = () => {
+//   const [state, dispatch] = useReducer(reducer, initialState);
 
-  return (
-    <>
-      <div>{state.count}</div>
-      <button
-        onClick={() =>
-          dispatch({
-            type: 'ADD_TODO',
-            payload: {
-              task: 'New Todo',
-              completed: false,
-              id: 2,
-            },
-          })
-        }
-      >
-        Add It!
-      </button>
-      <button onClick={() => dispatch({ type: 'COMPLETE_TODO' })}>
-        Complete It!
-      </button>
-    </>
-  );
-};
-
-export default ToDo;
+//   return (
+//     <>
+//       <div>{state.count}</div>
+//       <button
+//         onClick={() =>
+//           dispatch({
+//             type: 'ADD_TODO',
+//             payload: {
+//               task: 'New Todo',
+//               completed: false,
+//               id: 2,
+//             },
+//           })
+//         }
+//       >
+//         Add It!
+//       </button>
+//       <button onClick={() => dispatch({ type: 'COMPLETE_TODO' })}>
+//         Complete It!
+//       </button>
+//     </>
+//   );
+// };
